requirements.txt
requests==2.24.0
pyyaml==5.3.1
schedule==0.6.0